Finish reversing the sound subsystem
A lot of the mysteries of the Megadrive sound code are explained by
them being a subset of the implementation for the Amiga version, which
feels much more like the native version.

Some of the weirdness of the Amiga code seems to stem from the sound
data overlapping in memory with other data pulled in via overlays, I
think that's going to be the next thing to investigate.

A bunch of the audio data is referred to via offsets from a sound_base
memory location. I've written a little hacky thing to convert these
offsets into Ghidra scripting to create the labels.